+++ src/cmd/install/sorting.go
@@ -45,8 +45,6 @@ func sortToothFiles(toothFileList []toothfile.ToothFile) ([]toothfile.ToothFile,
 func visit(toothFile toothfile.ToothFile, toothFileMap map[string]toothfile.ToothFile, preVisited map[string]bool, visited map[string]bool, sorted *[]toothfile.ToothFile) error {
 	var err error
 
-	logger.Info("visiting", toothFile.Metadata().ToothPath)
-
 	if visited[toothFile.Metadata().ToothPath] {
 		return nil
 	}
@@ -56,8 +54,6 @@ func visit(toothFile toothfile.ToothFile, toothFileMap map[string]toothfile.Toot
 
 	preVisited[toothFile.Metadata().ToothPath] = true
 	for depToothPath := range toothFile.Metadata().Dependencies {
-		logger.Info("visiting dependency", depToothPath)
-
 		// Find the tooth file of the dependency.
 		dep, ok := toothFileMap[depToothPath]
 		if !ok {