+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `--available` flag for `lip show` command.
+
+### Fixed
+
+- Remove wrongly displayed debug information.
+
 ## [0.10.0] - 2023-02-12
 
 ### Added