+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for multiple GOPROXYs.
 - `--keep-possession` flag for `lip uninstall` command.
 - Automaic deletion of empty directories when uninstalling a tooth.
+- Support for file possession.
 
 ### Fixed