+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2023-02-17
+
 ### Added
 
 - `--available` flag for `lip show` command.
@@ -192,7 +194,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic functions: cache, install, list, show, tooth init, and uninstall.
 
-[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.10.0...HEAD
+[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/LiteLDev/Lip/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/LiteLDev/Lip/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/LiteLDev/Lip/compare/v0.8.3...v0.9.0
 [0.8.3]: https://github.com/LiteLDev/Lip/compare/v0.8.2...v0.8.3